Remove to \drop
