+++ tgbot/handlers/commands.py
@@ -2,11 +2,12 @@ from aiogram.types import BotCommand
 from aiogram import Bot
 
 bot_commands = [
-    BotCommand(command="start", description="🔄 Перезапустить бота."),
-    BotCommand(command="help", description="📒 Справка о командах."),
-    BotCommand(command="questions", description="❓ Вопросы."),
-    BotCommand(command="polls", description="📝 Опросы."),
-    BotCommand(command="answers", description="📩 Список ответов в ваших опросах."),
+    BotCommand(command="start", description="🔄 Перезапустить бота"),
+    BotCommand(command="help", description="📒 Справка о командах"),
+    BotCommand(command="questions", description="❓Вопросы"),
+    BotCommand(command="polls", description="📝 Опросы"),
+    BotCommand(command="answers", description="📩 Список ответов в ваших опросах"),
+    BotCommand(command="drop", description="🗑️ Сбросить сохраненные данные"),
 ]
