+++ tgbot/handlers/help.py
@@ -4,13 +4,13 @@ from aiogram.types import Message
 
 router = Router()
 help_text = """🤖 Список команд:\n\
-\n/start - 🔄 Перезапустить бота.\
-\n/help - 📒 Справка о командах.\
-\n/questions - ❓ Вопросы.\
-\n/polls - 📝 Опросы.  \
-\n/answers - 📩 Список ответов в ваших опросах. \
+\n/start - 🔄 Перезапустить бота\
+\n/help - 📒 Справка о командах\
+\n/questions - ❓ Вопросы\
+\n/polls - 📝 Опросы  \
+\n/answers - 📩 Список ответов в ваших опросах \
+\n/drop - 🗑️ Сбросить сохраненные данные \
 """
-
 # /help handler
 @router.message(Command("help"))
 async def cmd_help(message: Message):
